requirements.txt
streamlit
openai
pygsheets
pandas
heyoo
Pillow
python-dotenv